+++ server/api/orders.js
@@ -15,10 +15,14 @@ router.get('/', async (req, res, next) => {
 
 router.post('/:userId', async (req, res, next) => {
   try {
-    const order = await Order.create(req.body, {
-      include: [{ model: Drink }]
-    }).then(result => result.setUser(req.params.userId))
-    res.json(order)
+    if (req.user && req.user.isAdmin) {
+      const order = await Order.create(req.body, {
+        include: [{ model: Drink }]
+      }).then(result => result.setUser(req.params.userId))
+      res.json(order)
+    } else {
+      res.redirect('/')
+    }
   } catch (err) {
     next(err)
   }
@@ -27,10 +31,14 @@ router.post('/:userId', async (req, res, next) => {
 //  /:orderId routes
 router.get('/:orderId', async (req, res, next) => {
   try {
-    const order = await Order.findById(req.params.orderId, {
-      include: [{ model: Drink }]
-    })
-    res.json(order)
+    if (req.user && req.user.isAdmin) {
+      const order = await Order.findById(req.params.orderId, {
+        include: [{ model: Drink }]
+      })
+      res.json(order)
+    } else {
+      res.redirect('/')
+    }
   } catch (err) {
     next(err)
   }
@@ -38,12 +46,16 @@ router.get('/:orderId', async (req, res, next) => {
 
 router.put('/:orderId', async (req, res, next) => {
   try {
-    const [_, order] = await Order.update(req.body, {
-      returning: true,
-      where: { id: req.params.orderId },
-      include: [{ model: Drink }]
-    })
-    res.send(order)
+    if (req.user && req.user.isAdmin) {
+      const [_, order] = await Order.update(req.body, {
+        returning: true,
+        where: { id: req.params.orderId },
+        include: [{ model: Drink }]
+      })
+      res.send(order)
+    } else {
+      res.redirect('/')
+    }
   } catch (err) {
     next(err)
   }
@@ -51,12 +63,16 @@ router.put('/:orderId', async (req, res, next) => {
 
 router.delete('/:orderId', async (req, res, next) => {
   try {
-    await Order.destroy({
-      where: {
-        id: req.params.orderId
-      }
-    })
-    res.status(204).end()
+    if (req.user && req.user.isAdmin) {
+      await Order.destroy({
+        where: {
+          id: req.params.orderId
+        }
+      })
+      res.status(204).end()
+    } else {
+      res.redirect('/')
+    }
   } catch (err) {
     next(err)
   }
